requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
duckdb==1.2.2
exceptiongroup==1.3.0
fastapi==0.115.12
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-resources==6.4.5
jinja2==3.1.6
jiter==0.9.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.40.0
numpy==1.24.4
openai==1.79.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkgutil-resolve-name==1.3.10
plotly==6.1.2
protobuf==5.29.4
pyarrow==17.0.0
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.44.0
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
zipp==3.20.2
python-multipart
XlsxWriter==3.2.0
